Feat: 테스트 코드 추가

+++ src/main/java/com/project/food_ordering_service/global/dto/request/KakaoMapApiRequest.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class KakaoMapApiRequest {
 
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
 
 }
